Add support for generating a consumer for CSS properties that use <unicode-range-token> https://bugs.webkit.org/show_bug.cgi?id=289854
Reviewed by Darin Adler.

Moves support for consuming <unicode-range-token> values into its own
file as it is a primitive defined in css-syntax, not css-fonts.

This allow generation of:
    @font-face unicode-range

* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/css/CSSFontSelector.cpp:
* Source/WebCore/css/CSSProperties.json:
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Font.cpp:
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Font.h:
* Source/WebCore/css/parser/CSSPropertyParserConsumer+UnicodeRange.cpp: Added.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+UnicodeRange.h: Added.

Canonical link: https://commits.webkit.org/292235@main
