+++ Source/WebCore/css/parser/CSSPropertyParserConsumer+UnicodeRange.cpp
@@ -1,0 +1,165 @@
+/*
+ * Copyright (C) 2016 The Chromium Authors. All rights reserved
+ * Copyright (C) 2021 Metrological Group B.V.
+ * Copyright (C) 2021 Igalia S.L.
+ * Copyright (C) 2023-2024 Apple Inc. All rights reserved.
+ * Copyright (C) 2024-2025 Samuel Weinig <sam@webkit.org>
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ * 1. Redistributions of source code must retain the above copyright
+ *    notice, this list of conditions and the following disclaimer.
+ * 2. Redistributions in binary form must reproduce the above copyright
+ *    notice, this list of conditions and the following disclaimer in the
+ *    documentation and/or other materials provided with the distribution.
+ *
+ * THIS SOFTWARE IS PROVIDED BY APPLE INC. ``AS IS'' AND ANY
+ * EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
+ * PURPOSE ARE DISCLAIMED.  IN NO EVENT SHALL APPLE INC. OR
+ * CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+ * EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+ * PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
+ * OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+#include "config.h"
+#include "CSSPropertyParserConsumer+UnicodeRange.h"
+
+#include "CSSParserTokenRange.h"
+#include "CSSUnicodeRangeValue.h"
+#include <wtf/text/ParsingUtilities.h>
+#include <wtf/text/StringBuilder.h>
+#include <wtf/text/StringParsingBuffer.h>
+
+namespace WebCore {
+namespace CSSPropertyParserHelpers {
+
+static bool consumeOptionalDelimiter(CSSParserTokenRange& range, UChar value)
+{
+    if (!(range.peek().type() == DelimiterToken && range.peek().delimiter() == value))
+        return false;
+    range.consume();
+    return true;
+}
+
+static StringView consumeIdentifier(CSSParserTokenRange& range)
+{
+    if (range.peek().type() != IdentToken)
+        return { };
+    return range.consume().value();
+}
+
+static bool consumeAndAppendOptionalNumber(StringBuilder& builder, CSSParserTokenRange& range, CSSParserTokenType type = NumberToken)
+{
+    if (range.peek().type() != type)
+        return false;
+    auto originalText = range.consume().originalText();
+    if (originalText.isNull())
+        return false;
+    builder.append(originalText);
+    return true;
+}
+
+static bool consumeAndAppendOptionalDelimiter(StringBuilder& builder, CSSParserTokenRange& range, UChar value)
+{
+    if (!consumeOptionalDelimiter(range, value))
+        return false;
+    builder.append(value);
+    return true;
+}
+
+static void consumeAndAppendOptionalQuestionMarks(StringBuilder& builder, CSSParserTokenRange& range)
+{
+    while (consumeAndAppendOptionalDelimiter(builder, range, '?')) { }
+}
+
+static String consumeUnicodeRangeString(CSSParserTokenRange& range)
+{
+    if (!equalLettersIgnoringASCIICase(consumeIdentifier(range), "u"_s))
+        return { };
+    StringBuilder builder;
+    if (consumeAndAppendOptionalNumber(builder, range, DimensionToken))
+        consumeAndAppendOptionalQuestionMarks(builder, range);
+    else if (consumeAndAppendOptionalNumber(builder, range)) {
+        if (!(consumeAndAppendOptionalNumber(builder, range, DimensionToken) || consumeAndAppendOptionalNumber(builder, range)))
+            consumeAndAppendOptionalQuestionMarks(builder, range);
+    } else if (consumeOptionalDelimiter(range, '+')) {
+        builder.append('+');
+        if (auto identifier = consumeIdentifier(range); !identifier.isNull())
+            builder.append(identifier);
+        else if (!consumeAndAppendOptionalDelimiter(builder, range, '?'))
+            return { };
+        consumeAndAppendOptionalQuestionMarks(builder, range);
+    } else
+        return { };
+    return builder.toString();
+}
+
+struct UnicodeRange {
+    char32_t start;
+    char32_t end;
+};
+
+// MARK: <unicode-range-token> consuming (unresolved)
+static std::optional<UnicodeRange> consumeUnicodeRangeTokenUnresolved(CSSParserTokenRange& range)
+{
+    return readCharactersForParsing(consumeUnicodeRangeString(range), [&](auto buffer) -> std::optional<UnicodeRange> {
+        if (!skipExactly(buffer, '+'))
+            return std::nullopt;
+        char32_t start = 0;
+        unsigned hexDigitCount = 0;
+        while (buffer.hasCharactersRemaining() && isASCIIHexDigit(*buffer)) {
+            if (++hexDigitCount > 6)
+                return std::nullopt;
+            start <<= 4;
+            start |= toASCIIHexValue(*buffer++);
+        }
+        auto end = start;
+        while (skipExactly(buffer, '?')) {
+            if (++hexDigitCount > 6)
+                return std::nullopt;
+            start <<= 4;
+            end <<= 4;
+            end |= 0xF;
+        }
+        if (!hexDigitCount)
+            return std::nullopt;
+        if (start == end && buffer.hasCharactersRemaining()) {
+            if (!skipExactly(buffer, '-'))
+                return std::nullopt;
+            end = 0;
+            hexDigitCount = 0;
+            while (buffer.hasCharactersRemaining() && isASCIIHexDigit(*buffer)) {
+                if (++hexDigitCount > 6)
+                    return std::nullopt;
+                end <<= 4;
+                end |= toASCIIHexValue(*buffer++);
+            }
+            if (!hexDigitCount)
+                return std::nullopt;
+        }
+        if (buffer.hasCharactersRemaining())
+            return std::nullopt;
+        return { { start, end } };
+    });
+}
+
+RefPtr<CSSValue> consumeUnicodeRangeToken(CSSParserTokenRange& range)
+{
+    auto rangeCopy = range;
+
+    auto unicodeRange = consumeUnicodeRangeTokenUnresolved(rangeCopy);
+    rangeCopy.consumeWhitespace();
+    if (!unicodeRange || unicodeRange->end > UCHAR_MAX_VALUE || unicodeRange->start > unicodeRange->end)
+        return nullptr;
+
+    range = rangeCopy;
+    return CSSUnicodeRangeValue::create(unicodeRange->start, unicodeRange->end);
+}
+
+} // namespace CSSPropertyParserHelpers
+} // namespace WebCore
